*** empty log message ***
svn path=/trunk/boinc/; revision=1689

+++ apps/upper_case.C
@@ -2,18 +2,18 @@
 // Version 1.0 (the "License"); you may not use this file except in
 // compliance with the License. You may obtain a copy of the License at
 // http://boinc.berkeley.edu/license_1.0.txt
-// 
+//
 // Software distributed under the License is distributed on an "AS IS"
 // basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
 // License for the specific language governing rights and limitations
-// under the License. 
-// 
-// The Original Code is the Berkeley Open Infrastructure for Network Computing. 
-// 
+// under the License.
+//
+// The Original Code is the Berkeley Open Infrastructure for Network Computing.
+//
 // The Initial Developer of the Original Code is the SETI@home project.
 // Portions created by the SETI@home project are Copyright (C) 2002
-// University of California at Berkeley. All Rights Reserved. 
-// 
+// University of California at Berkeley. All Rights Reserved.
+//
 // Contributor(s):
 //
 
@@ -103,7 +103,7 @@ int do_checkpoint(MFILE& mf, int nchars) {
 
     if (cpu_time) {
         double cur_cpu;
-        int cur_mem;
+        double cur_mem;
         // print our own information about cpu time
         fprintf(app_time, "%f\n", difftime(time(0), my_start_time));
         fflush(app_time);
@@ -240,7 +240,7 @@ int main(int argc, char **argv) {
     if (random_exit) exit(-10);
     fprintf(stderr, "APP: upper_case ending, wrote %d chars\n", nchars);
     double cur_cpu;
-    int cur_mem;
+    double cur_mem;
     boinc_cpu_time(cur_cpu, cur_mem);
     time_file.printf("%f\n", cur_cpu);
     time_file.flush();
